display error when the aggregator bearer token is rejected
